+++ src/alert/alert.tsx
@@ -4,6 +4,7 @@ import classNames from 'classnames';
 import exceptionIcon from '@jetbrains/icons/exception';
 import checkmarkIcon from '@jetbrains/icons/checkmark';
 import warningIcon from '@jetbrains/icons/warning';
+import infoIcon from '@jetbrains/icons/info-filled';
 import closeIcon from '@jetbrains/icons/close';
 
 import Icon, {Color} from '../icon/icon';
@@ -13,6 +14,8 @@ import dataTests from '../global/data-tests';
 import Button from '../button/button';
 import Theme, {ThemeProvider} from '../global/theme';
 import {type AlertProps, AlertType} from './alert.interface';
+import AlertHeading from './alert-heading';
+import AlertActions from './alert-actions';
 
 import styles from './alert.css';
 
@@ -30,6 +33,7 @@ const TypeToIcon: Partial<Record<AlertType, string>> = {
   [AlertType.ERROR]: exceptionIcon,
   [AlertType.SUCCESS]: checkmarkIcon,
   [AlertType.WARNING]: warningIcon,
+  [AlertType.INFO]: infoIcon,
 };
 
 /**
@@ -39,7 +43,8 @@ const TypeToIcon: Partial<Record<AlertType, string>> = {
 const TypeToIconColor: Partial<Record<AlertType, Color>> = {
   [AlertType.ERROR]: Color.RED,
   [AlertType.SUCCESS]: Color.GREEN,
-  [AlertType.WARNING]: Color.WHITE,
+  [AlertType.WARNING]: Color.ORANGE,
+  [AlertType.INFO]: Color.BLUE,
 };
 
 interface State {
@@ -93,6 +98,8 @@ export default class Alert extends PureComponent<AlertProps, State> {
   hideTimeout?: number;
 
   static Type = AlertType;
+  static Heading = AlertHeading;
+  static Actions = AlertActions;
 
   closeRequest = (event: React.MouseEvent<HTMLElement>) => {
     this.startCloseAnimation();
@@ -127,9 +134,7 @@ export default class Alert extends PureComponent<AlertProps, State> {
   private _getCaption() {
     return (
       <span
-        className={classNames(styles.caption, this.props.captionClassName, {
-          [styles.withCloseButton]: this.props.closeable,
-        })}
+        className={classNames(styles.caption, this.props.captionClassName)}
         onClick={this._handleCaptionsLinksClick}
         // We only process clicks on `a` elements, see above
         role='presentation'
@@ -147,7 +152,7 @@ export default class Alert extends PureComponent<AlertProps, State> {
     const glyph = TypeToIcon[this.props.type];
 
     if (glyph) {
-      return <Icon glyph={glyph} className={styles.icon} color={TypeToIconColor[this.props.type] || Color.DEFAULT} />;
+      return <Icon glyph={glyph} color={TypeToIconColor[this.props.type] || Color.DEFAULT} />;
     }
     if (this.props.type === AlertType.LOADING) {
       return <Loader className={styles.loader} />;
@@ -171,6 +176,7 @@ export default class Alert extends PureComponent<AlertProps, State> {
       className,
       'data-test': dataTest,
       theme,
+      afterMessage,
     } = this.props;
 
     const classes = classNames(className, {
@@ -196,10 +202,11 @@ export default class Alert extends PureComponent<AlertProps, State> {
       >
         {this._getIcon()}
         {this._getCaption()}
+        {afterMessage}
         {this.props.closeable ? (
           <Button
             icon={closeIcon}
-            className={classNames(styles.close, closeButtonClassName)}
+            className={closeButtonClassName ?? undefined}
             data-test='alert-close'
             aria-label='close alert'
             onClick={this.closeRequest}