+++ src/icon/icon.constants.ts
@@ -10,6 +10,7 @@ export enum Color {
   MAGENTA = 'magenta',
   RED = 'red',
   WHITE = 'white',
+  ORANGE = 'orange',
 }
 
 /**